Separate mas field from merch card
