+++ libs/utils/decorate.js
@@ -55,7 +55,8 @@ export function decorateButtons(el, size) {
   if (buttons.length === 0) return;
 
   buttons.forEach((button) => {
-    if (button.classList.contains('merch-card-autoblock') && button.classList.contains('link-block')) return;
+    if ((button.classList.contains('merch-card-autoblock') || button.classList.contains('merch'))
+      && button.classList.contains('link-block')) return;
     const parent = button.parentElement;
     if (shouldBlockFreeTrialLinks(button)) return;
     let target = button;
